skills/kanban-implementer/core/pyproject.toml
[project]
name = "kanban-implementer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
kanban-select = "kanban_implementer.__main__:main_select"
kanban-update = "kanban_implementer.__main__:main_update"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.ruff]
line-length = 100

[tool.mypy]
strict = true
